AtomData: class template -> class; h/cpp splits
